Refactors postprocessing code
Rearranges some modules to minimize dependencies for postprocessing
Rewrites some code to minimize dependencies
Adds code to parse out postal codes
Small bugfixes
