+++ postprocessing.py
@@ -1,34 +1,66 @@
 """
 This model applies different post processing steps on the results file of the detectron2 model
 """
+import argparse
 import csv
+import json
+import os
 from copy import deepcopy
 from datetime import date, datetime
 from pathlib import Path
-from typing import List
+from typing import Any, List
 
 import geopy.distance
 import numpy as np
+import numpy.typing as npt
 import pycocotools.mask as mask_util
-import torch
 from panorama.client import PanoramaClient
 from shapely.geometry import LineString, Point
+from shapely.ops import nearest_points
 from tqdm import tqdm
-from triangulation.helpers import (
-    get_panos_from_points_of_interest,
-)  # pylint: disable-all
+from triangulation.helpers import get_panos_from_points_of_interest  # pylint: disable-all
 from triangulation.masking import get_side_view_of_pano
 from triangulation.triangulate import triangulate
 
-from utils import (
-    calculate_distance_in_meters,
-    get_bridge_information,
-    get_container_locations,
-    get_permit_locations,
-    save_json_data,
-    write_to_csv,
-)
 from visualizations.stats import DataStatistics
+from visualizations.utils import get_bridge_information, get_permit_locations
+
+
+def calculate_distance_in_meters(line: LineString, point: Point) -> float:
+    """
+    Calculates the shortest distance between a line and point, returns a float in meters
+    """
+    closest_point = nearest_points(line, point)[0]
+    return float(geopy.distance.distance(closest_point.coords, point.coords).meters)
+
+
+def get_container_locations(file: Path) -> List[List[float]]:
+    """
+    Returns locations of containers from a csv file in lat, lon order
+    """
+    container_locations = []
+    with open(file) as csv_file:
+        csv_reader = csv.reader(csv_file, delimiter=",")
+        next(csv_reader)  # skip first line
+        for row in csv_reader:
+            container_locations.append([float(row[0]), float(row[1])])
+    return container_locations
+
+def save_json_data(data: Any, filename: Path, output_folder: Path) -> None:
+    """
+    Write the data to a json file
+    """
+    with open(output_folder / filename, "w") as f:
+        json.dump(data, f)
+
+
+def write_to_csv(
+    data: npt.NDArray[npt.NDArray[Any]], header: List[str], filename: Path
+) -> None:
+    """
+    Writes a list of list with data to a csv file.
+    """
+    np.savetxt(filename, data, header=",".join(header), fmt="%d", delimiter=",", comments="")
 
 
 class PostProcessing:
@@ -89,30 +121,24 @@ class PostProcessing:
         """
         predictions_to_keep = []
         print("Filtering based on angle")
-        for prediction in tqdm(self.stats.data):
+        running_in_k8s = "KUBERNETES_SERVICE_HOST" in os.environ
+        for prediction in tqdm(self.stats.data, disable=running_in_k8s):
             response = PanoramaClient.get_panorama(
                 prediction["pano_id"].rsplit(".", 1)[0]
             )  # TODO: Not query, look at the database!
             heading = response.heading
             height, width = prediction["segmentation"]["size"]
-            mask = torch.from_numpy(
-                get_side_view_of_pano(width, height, heading, self.mask_degrees)
-            )[
-                :, :, 0
-            ]  # mask is 3D
-            mask = (
-                1 - mask.float()
-            )  # Inverse mask to have region that we would like to keep
-            segmentation_mask = torch.from_numpy(
-                mask_util.decode(prediction["segmentation"])
-            ).float()
+            mask = get_side_view_of_pano(width, height, heading, self.mask_degrees)[:, :, 0]
+            # Inverse mask to have region that we would like to keep
+            mask = 1 - mask
+            segmentation_mask = mask_util.decode(prediction["segmentation"])  # .float()
             overlap = segmentation_mask * mask
             # prediction fully in side view
-            if torch.sum(overlap) == torch.sum(segmentation_mask):
+            if np.sum(overlap) == np.sum(segmentation_mask):
                 predictions_to_keep.append(prediction)
             # prediction partially in side view
             # Keep predictions if minimal 50% is in the side view
-            elif torch.sum(overlap) / torch.sum(segmentation_mask) > 0.5:
+            elif np.sum(overlap) / np.sum(segmentation_mask) > 0.5:
                 predictions_to_keep.append(prediction)
         print(
             f"{len(self.stats.data) - len(predictions_to_keep)} out of the {len(self.stats.data)} are filtered based on the angle"
@@ -192,29 +218,38 @@ class PostProcessing:
         sorted_scores = np.array(scores)[sorted_indices]
 
         write_to_csv(
-            [
+            np.array([
                 prioritized_containers[:, 0],
                 prioritized_containers[:, 1],
                 sorted_scores,
                 permit_distances_sorted,
                 bridges_distances_sorted,
-            ],
+            ]),
             ["lat", "lon", "score", "permit_distance", "bridge_distance"],
             self.output_folder / self.prioritized_file,
         )
 
 
 if __name__ == "__main__":
-    output_path = Path("Test")
+    parser = argparse.ArgumentParser(description="Run postprocessing for container detection pipeline")
+    parser.add_argument("input_path", type=Path, help="Full path to input file")
+    parser.add_argument("output_path", type=Path, help="Full path to output dir")
+    parser.add_argument("permits_file", type=Path, help="Full path to permits file")
+    parser.add_argument("bridges_file", type=Path, help="Full path to bridges file")
+    args = parser.parse_args()
+
     postprocess = PostProcessing(
-        Path("coco_instances_results_14ckpt.json"), output_folder=output_path
+        args.input_path,
+        output_folder=args.output_path,
+        permits_file=args.permits_file,
+        bridges_file=args.bridges_file,
     )
     postprocess.filter_by_size()
     postprocess.filter_by_angle()
     postprocess.find_points_of_interest()
     panoramas = get_panos_from_points_of_interest(
-        output_path / "points_of_interest.csv", date(2021, 3, 18), date(2021, 3, 17)
+        args.output_path / "points_of_interest.csv", date(2021, 3, 18), date(2021, 3, 17)
     )
     postprocess.prioritize_notifications()
     for pano in panoramas:
-        PanoramaClient.download_image(pano, output_location=output_path)
+        PanoramaClient.download_image(pano, output_location=args.output_path)
